[~] TODO(Maksim Mishenko): delete after merge
