+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/core/ui/screen_with_controls/TimerSlideShow.kt
@@ -31,7 +31,7 @@ class TimerSlideShow @Inject constructor() {
                 var currentTime = System.currentTimeMillis()
                 while (isActive && currentTime - startTime < period * 1000)
                     currentTime = System.currentTimeMillis()
-                //TODO(Stanislav Radchenko): @Maksim Mishenko improve this code
+                //TODO(Stanislav Radchenko): @Artem Gruzdev improve this code
                 if (isActive && isPlay)
                     callbackToEnd()
             }
