+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/screen/message/component/ManyMessagesScreen.kt
@@ -1,17 +1,15 @@
 package band.effective.office.tv.screen.message.component
 
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
 import band.effective.office.tv.domain.model.message.BotMessage
+import band.effective.office.tv.screen.components.ProgressIndicator
 import coil.ImageLoader
 
 @Composable
@@ -46,36 +44,6 @@ fun ManyMessagesScreen(
                 onClickBackButton = onClickButton,
                 textColor = textColor
             )
-        }
-    }
-}
-
-
-// TODO(Maksim Mishenko): delete after merge
-@Composable
-fun ProgressIndicator(
-    modifier: Modifier,
-    elementModifier: Modifier,
-    count: Int,
-    currentIndex: Int,
-    elementColor: Color = MaterialTheme.colors.secondary,
-    currentElementColor: Color = Color.White
-) {
-    Row(modifier = modifier, horizontalArrangement = Arrangement.Center) {
-        for (index in 0 until count) {
-            val animatedColor =
-                animateColorAsState(
-                    if (index == currentIndex) currentElementColor else elementColor
-                )
-            Row {
-                Box(
-                    modifier = elementModifier
-                        .width(8.dp)
-                        .height(8.dp)
-                        .background(animatedColor.value)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-            }
         }
     }
 }
